Move ShouldCpuAllocatorUseArena to allocator_utils.h/cc.
