+++ onnxruntime/core/framework/allocator_utils.cc
@@ -8,6 +8,8 @@
 #include <sstream>
 #include <unordered_map>
 
+#include <absl/base/config.h>
+
 #include "core/common/logging/logging.h"
 #include "core/common/narrow.h"
 #include "core/framework/bfc_arena.h"
@@ -73,6 +75,23 @@ AllocatorPtr CreateAllocator(const AllocatorCreationInfo& info) {
   } else {
     return device_allocator;
   }
+}
+
+bool ShouldCpuAllocatorUseArena([[maybe_unused]] bool is_arena_requested) {
+#if defined(USE_JEMALLOC) || defined(USE_MIMALLOC)
+  // We use these allocators instead of the arena.
+  return false;
+#elif defined(ABSL_HAVE_ADDRESS_SANITIZER)
+  // Using the arena may hide memory issues. Disable it in an ASan build.
+  return false;
+#else
+  // Disable the arena for 32-bit builds because it may run into an infinite loop on integer overflow.
+  if constexpr (sizeof(void*) == 4) {
+    return false;
+  } else {
+    return is_arena_requested;
+  }
+#endif
 }
 
 }  // namespace onnxruntime
