+++ onnxruntime/core/providers/cpu/cpu_execution_provider.cc
@@ -3,7 +3,7 @@
 
 #include "core/providers/cpu/cpu_execution_provider.h"
 
-#include "core/framework/cpu_allocator_utils.h"
+#include "core/framework/allocator_utils.h"
 #include "core/framework/op_kernel.h"
 #include "core/framework/kernel_registry.h"
 #include "core/framework/int4.h"
