+++ onnxruntime/core/session/environment.cc
@@ -4,7 +4,6 @@
 #include "core/session/environment.h"
 #include "core/session/allocator_adapters.h"
 #include "core/framework/allocator_utils.h"
-#include "core/framework/cpu_allocator_utils.h"
 #include "core/graph/constants.h"
 #include "core/graph/op.h"
